Improved parsing of speaker citations
We now attempt to look up the speaker on google scholar with their
affiliation to avoid potentially finding an unrelated person with the
same name. If we do not find a person with that name and affiliation, we
just look up the name.

Thank you for alerting us to this issue, @bangerth! Closes #3.