Improved fix_order function
The fix_order function was causing problems where it would move
multipart minisymposia to improve the gumband penalty without regard for
whether that was a valid timeslot to move them to. In the future, we may
wish to consider whether moving the minisymposia causes a conflict where
the speaker is oversubscribed, but that will increase the cost of the
function.